Fix zip cuffs size (#9011)
## What this PR does
Adds a size property for box of zip cuffs with a value of MEDIUM
(normal)
## Why this is good for the game
Adds atmosphere and RP playability to Squad Leader, such as in
situations with surviving colonists.
## Change Images

![image](https://github.com/user-attachments/assets/cbca0e25-b021-4d2f-9b36-2566523144d0)
## Testing
On a local server. Interacting with: B12 pattern marine armor, G8-A
general, M11 pattern marine helmet
## Changelog

:cl:
fix: added size property for box of zip cuffs with value MEDIUM (normal)
/:cl:

Co-authored-by: urigrachev <urigrachev@team.amocrm.com>