Small non-underground area in the LV lake house (#8879)
# About the pull request

This change adds a small 5x5 area in the LV north central caves. It has
no roof, and in the middle of it is a broken comms tower.

This is my first proper PR for byond, SS13, and of course, first time
doing anything map-related. It seemed to work well when I was testing
it, though. I would be open to ideas about what else to place in this
little area, as I was unsure. It is meant to be a broken comms relay of
sorts.

# Explain why it's good for the game

The reason why I am making this change is because I have noticed that
the LV is the only map in the game where the sensor tower is neither in
an open area nor close to one. It's pretty deep. I have also enjoyed
non-underground areas very deep in the map quite a bit in other maps
such as Hybrisa. It shakes things up a bit honestly.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.
<img width="471" alt="Screenshot 2025-03-26 at 21 45 42"
src="https://github.com/user-attachments/assets/05466bf0-93ff-41f8-bd2a-a40ed5e3e6f7"
/>

</details>


# Changelog
:cl:
add: Addeded a new area to LV lake house area that isn't underground.
/:cl: