Replaces the pred AI announcement with an Elder Overseer  (#8779)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
Replace the Predator AI announcement with an Elder Overseer that the
Predator Council and admins may use to communicate with living
predators.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
It's better than the horrible call of duty nuke AI announcement.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->

<!-- !! If you are modifying sprites, you **must** include one or more
in-game screenshots or videos of the new sprites. !! -->

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/908e2f80-4d5e-41d5-bc15-8102936d2557


https://github.com/user-attachments/assets/7b4b3c30-be6d-4471-b378-f040da341d47



https://github.com/user-attachments/assets/3878e23a-120d-4e00-8a04-8fca34098457





</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl: Joe Lampost, Forest2001
add: Changes the Ai announcement to an elder overseer that only: Yautja,
thralls and hellhounds can see
add: Adds a button for yautja council members to message all alive preds
yautja.
del: Removed the Yautja AI announcement 
fix: The youngblood terminator now displays what it is in the text input
soundadd: added four sounds for the elder overseer
code: moves the whitelist panel to a new tab in the OOC panel
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>