Adds laser designator targetting to mortar (#8725)
# About the pull request

Adds an aiming toggle to the mortar, which lets users link a laser
designator, and fire directly on that laser designators lases as an
alternative to firing on coordinates.

# Explain why it's good for the game

This addition gives mortar operators a a second choice on how they can
fire on a target in order to give them more flexibility.

"Lase mode", toggleable with Alt+Clicking on the mortar, lets mortar
operators link any laser designator to the mortar. The mortar then will
automatically aim at any target lased by the laser designator after a
short delay, just like for CAS targets. The benefits of this mode is
that the mortar operator can fire on targets much faster, and the mortar
shells have 100% accuracy in this mode. The drawbacks compared to
coordinate mode, is that a lase has to be active for the mortar operator
to fire on the target. The mortar operator also cannot offset the
mortar, so it is strictly direct fire. If the laser designator stops the
lasing, the mortar will automatically reset, unable to fire until a
laser target is acquired again. The mortar operator cannot lase and try
to fire the mortar at the same time, so they will need to find a buddy
that can lase targets for them.

The aim is for the "laser designator targetting" mode to not be a direct
upgrade or downgrade to coordinate firing mode, but a sidegrade, where
the mortar operator can switch modes if needed.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/490fb4a5-4edd-4d02-a236-7658b7f4edba

</details>


# Changelog
:cl: Mikrel0712
add: Added a laser targetting mode to the M402 Mortar. You can now link
a laser designator to the mortar, and fire directly on laser targets as
an alternative to firing on coordinates.
/:cl: