CLF / CO Surv fix (#9056)
# About the pull request

Gives CLF survivor landmarks CO_equipment overrides that are their own
default equipment

# Explain why it's good for the game

If someone rolls for CO surv on NV (lol) and the CLF insert fires, he
will be spawned as a regular CO surv in the insert, which is
undesirable. This fixes that.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: CO survs on NV will no longer be UASF COs in the CLF insert
/:cl: