Autowiki Tweaks. Skips/Armour damage table 0s/Xeno strain comparison (#9078)
# About the pull request

Skips some things like bursting pipes which are a... subtype of grenade.

SG barrels? These aren't removable it's not really an attachment ugh.
88 Burst Stock ??

Other barrels and stocks that are basically just the weapon because
they're not removable.

14 different kinds of bayonet which are basically the same. The urge to
add some sort of generic bayonet for just autowiki is high but my effort
is low. Cause I'd need to add it to every gun that has bayonets.

Armour damage table thingie uses isnull to allow 0 values in the table.

Implement Drathek's thingie for looking at the differences for strains
and base. Though just using a loop and adding colour.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Uhhhh wiki shouldn't be flooded with bayonets, and random attachies. And
grenades that pretend they're bursting pipes. Armour damage tables
shouldn't say armour-40 when it should be 0.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas, Drathek
qol: Autowiki: Ignores most bayonets, and other misc. attachments.
Compares difference in base vs strain stats. Switches to using isnull so
armour damage table populates 0s.
/:cl: