Removes some large magazine pouches on Hybrisia. (#9038)
# About the pull request

This removes some of the large magazine pouches on Hybrisia. It's not a
good state for a map to be in when you can loot literally any random
goon body as a marine and get a free large magazine pouch, ESPECIALLY
when survivors spawn with them in certain presets already.

# Explain why it's good for the game

Large magazine pouches should not be obtainable from just looting random
bodies that spawn on the map in such high volumes, especially not as
part of your roundstart loadout.

# Testing Photographs and Procedure
N/A


# Changelog

:cl:
balance: Security on Hybrisia Prospera have had buget cuts and no longer
have large magazine pouches.
/:cl: