Underbarrel Extinguisher Expansion (#9041)
# About the pull request

This is the same PR that popped up last week, I just tested every gun to
make sure it actually works.

Expands the variety of guns that you can attach the underbarrel
extinguisher to. You can now attach it to guns such as the m39, M4RA,
breaching shottie, mou, and scout gun (if you're feeling spicy). You
still can't attach it to guns where it wouldn't make a lot of sense, so
no attaching it to pistols, revolvers, bolt action rifles, lever
actions, etc.

# Explain why it's good for the game

The underbarrel extinguisher is a useful little thing, and it doesn't
make a ton of sense why you can't stick it on a submachine gun. Tempted
to name this quality of life, but tagging this balance just in case.
Hopefully shouldn't affect game balance beyond clearing up a few
headaches.

# Testing Photographs and Procedure


<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/17ba3850-981c-4efb-a77c-ee1f0f62db2b)


![image](https://github.com/user-attachments/assets/ce71a30c-48a3-49a8-aeb7-e6da12fc0d1d)


![image](https://github.com/user-attachments/assets/194f5106-ab10-4176-b688-bb08299cc17f)

</details>


# Changelog

:cl:
balance: Underbarrel extinguishers can now be attached to the m39, M4RA,
mou, breaching shotty, and others.
qol:
/:cl: