Autowiki: Gun templates now omit unwielded stats for two handers (#9091)
# About the pull request

This PR makes it so unwielded gun stats for the autowiki are overrided
to "N/A" for weapons that are two handers (the in game UI already omits
this data).

I also made it so the new xeno comparisons round to the nearest 0.01
because vampire wanted to round special.

![image](https://github.com/user-attachments/assets/db6baaf6-4c6e-48f6-9729-3be6f9d6acb5)

# Explain why it's good for the game

Less confusion for autowiki gun templates.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/01a4053d-005d-403c-9fa5-7e5725752bfe)

[autowiki_edits.txt](https://github.com/user-attachments/files/19811666/autowiki_edits.txt)

</details>

# Changelog
:cl: Drathek
code: Autowiki gun templates now lists unwielded stats as N/A for two
handers
code: Autowiki xeno comparisons now round to the nearest hundredth
/:cl: