incend mag removal for CO (#9101)
# About the pull request

It removes the 40 round incendiary MK2 magazines from the CO 

# Explain why it's good for the game

The Commanding Officer should be better equipped than an average marine,
but not by this caliber. The Commanding Officer gets the best armor, the
best weapon (MK1 with IFF), the best sidearm (with stuns that can kill a
lurker with 2 clicks if done right). They don't need the MK2 incendiary
mags. It's also very overpowered to say the least.



# Changelog

:cl: stalkerino
balance: removed 40 round incendiary mags from CO
/:cl: