Obliterates thermal glasses from in-rotation maps and goggle spawners. (#9066)
# About the pull request

Removes thermal goggles from random chance spawners and preset spawns on
every map that I have found to have them.

# Explain why it's good for the game
Survivor, over the years, has progressively become stronger and stronger
and stronger. This is an extremely small step, although it is one of the
most egregious examples of overpowered map loot available to both
survivors and marines that does not fit the current version of CM.

Having permanent, full awareness of every single xenomorphs position
using binoculars and being able to relay it is extremely unhealthy for
the game, and is bar nothing one of the single most powerful things in
the game - and, frankly, it shouldn't be available at all to a role that
is designed to survive rather than to be a doom frag 9000 role, and much
less should it fall into marine hands such as being given to the scout
specialist or demolitionist.


# Testing Photographs and Procedure
N/A.


# Changelog

:cl:
balance: Removes thermals as a possible selection from random goggle
spawners such as on Fiorina and from pre-selected map spawns on Sorokyne
Strata and Kutjevo.
/:cl:

Co-authored-by: Git-Nivrak <59925169+Git-Nivrak@users.noreply.github.com>