Project ARES: Turret IFF Controls (#8226)
# About the pull request

As title, gives staff ability to change the IFF settings of the AI Core
turrets. They will never fire on Working Joes.

# Explain why it's good for the game

Makes it easier for staff to do things in regards to solo holds in the
AI Core and other rarities, plus event options.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added the ability for staff to change IFF settings on the AI Core
turrets via remote interface.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>