thunderdome cleanup system (#9122)
# About the pull request

adds an admin button to the events panel to clean the thunderdome.
Allows you to select whether to delete the mobs inside, and resets the
turfs by reloading the tdome as a template.

I used tg's thunderdome cleanup as a base for this PR, I couldn't find
the PR that originally added the system but this one was helpful.
https://github.com/tgstation/tgstation/pull/65545

# Explain why it's good for the game

I've had multiple staff independently come request this from me; ease of
clearing up loose stuff in the tdome without spamming bombs which just
asks for trouble.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/eec0bd37-2e56-4ad1-875b-21e4a67cde60

</details>


# Changelog

:cl:
admin: added a new button to reset the tdome
/:cl: