Fix/UI: Adds a scrollbar to individual security records (#8984)
# About the pull request
Actually just kills one scrollbar and moves it to a section to encompass
everything. Also centers and uses a box for the "Security Records"
header because of auto header title height *cry
And centers the new record button whilst I'm at it.
And also moves the photo window so it's not touching the side.
Also also spaces the print and delete button on opposite ends instead of
next to each other.
And Also And aligns the search bar and the squad panel with the m=0
stuff.
Resolves: #8981
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fixes a thing... Kinda
Looks nicer centered? I dunno maybe that's subjective.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

<details>
<summary>Bonus OW Alignment Thingie</summary>

Before:
You can barely see that the search bar is wider on the left by like one
pixel.

![Screenshot 2025-04-10
132635](https://github.com/user-attachments/assets/8d49c8ff-1e03-4f32-816d-870e1e852859)

After:

![Screenshot 2025-04-10
132229](https://github.com/user-attachments/assets/b049a916-e64d-47c4-927f-379116675c7f)


</details>

Before:

![Screenshot 2025-04-05
064431](https://github.com/user-attachments/assets/2b42dc74-46d6-43c2-9a7d-09dcbe532daf)

After:

![Screenshot 2025-04-05
070614](https://github.com/user-attachments/assets/4f779b3f-967f-4bbf-b4ce-8b2246e81fe2)

Before:

![Screenshot 2025-04-05
064414](https://github.com/user-attachments/assets/b177f30b-9af6-4ccb-9fbc-5707e03ba487)

After:

![Screenshot 2025-04-05
070618](https://github.com/user-attachments/assets/58500b3b-2352-4884-94bd-c365ac240c37)

Before:


![image](https://github.com/user-attachments/assets/037a9211-5c52-43d2-b199-dd326a355cfa)

After:


![image](https://github.com/user-attachments/assets/b8bbcef6-7220-495e-9fed-cbe2ba5211dc)

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>




# Changelog
:cl: MistChristmas
ui: Security Records: adds a scrollbar to Individual records. Centers
Title and New Record button. Slightly moves the Photo window. Moves the
record delete button to the right hand-side. OW Console: Slightly aligns
the search bar and Squad Panel in the Squad Monitor.
/:cl: