Banished can now alt click to exit a tunnel (#9176)
# About the pull request

This PR fixes some confusion on an edge case if you get banished while
in a tunnel. Prior to this change: when banished, alt click and ctrl
clicking on a tunnel is disabled. Alt clicking a tunnel is a way to exit
a tunnel. Even though there's a separate verb that would allow you to
exit the tunnel in this situation (Exit-Tunnel), now alt click will
still work to exit the tunnel.

# Explain why it's good for the game

Less confusion for the banished player I guess? E.g.
https://github.com/cmss13-devs/cmss13/issues/9097#issuecomment-2819753196

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/_A94bJBXok0

</details>


# Changelog
:cl: Drathek
qol: Banished xenos in a tunnel can now exit using alt+click (they could
have used the exit-tunnel verb however)
/:cl: