adds cooldown to makeup application (#9213)
# About the pull request

fixes: #9212 by adding a cooldown to applying makeup

changed one letter vars while there.

# Explain why it's good for the game

makeup takes time and precision to apply and shouldn't be able to be
spammed on anyone, and we shouldn't allow users to spam each other with
popups that take screen priority.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog


:cl:
fix:makeup application can no longer be spammed
/:cl: