Fix: Auto verbing for '!' and '? ' (#9179)
# About the pull request
Add the verb argument to the speech_problems thing.
Uses the verbs found in the languages. Also makes them lists so you can
add to them for extra flavour. And allows English to still have it's
three choices for exclaiming

This affects **local**, and **radio**.

Changes some language verbs to English ones as it's weird to have them
not English, when you can understand at least that they're shouting in
your face or trying to ask you a question

Also allows same faction mobs to know the speaker over the radio if
they're speaking a language they don't understand.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fix a thing that has been broken for... however long.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Messages ending in '!' and '?' change the verb from 'says' to
'shouts/yells/excalims' or 'asks' again or their language equivalent.
/:cl: