Accessory fixes & Tactical masks can be worn as accessories (#8894)
# About the pull request

Title. Allows the wearing of scarves and other masks on your armor or
uniform through use item interaction.
Thank you CM Dev.

# Explain why it's good for the game

Adds the functionality to make masks and other clothing items such as
hats to be treated as accessories without refactoring the entirety of
accessory code.

Currently, you need to use it while in your hand to change it from and
to an accessory.

By god, I am not refactoring accessory code.

This means you can smoke something while looking cool or wear another
mask under your mask, I guess.

Prevents accessories from attaching to uniforms while its in a storage
item as an alternative of #9110
closes #9035

# Testing Photographs and Procedure

![image](https://github.com/user-attachments/assets/0076e0d2-5f66-4f2a-a544-a1028d9c42e9)



# Changelog
:cl:
add: Scarves and some masks can be used as accessory items! Use them
inhand to change them back and to an accessory.
fix: Armbands no longer hide when sleeves are cut or jacket is removed
fix: Accessories may no longer be attached while it is in a storage
item.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>