lungemines now gib the user (#9181)
# About the pull request

my main goal is to blow up
lunge mines will now gib the user on hit, this is to make it more in
line with the fact that for most marine roles it will gib. i considered
making it os that the user blows up with the same severity as the hit so
if they have extremely heavy armour (B18 or similar) they can survive
but decided to simply call gib although i can easily change it

damaged lunge mines do not gib the user, they deal relatively low damage
in comparison

# Explain why it's good for the game

i don't find lunge mines to be bad for the game in their current state,
BUT i think it's bad that you can use it and sometimes entirely avoid
damage or die and be revived anyway, this item is niche as well.

for further context, i think that it's fine for marines to charge into
suicidal situations or OD/PB themselves to prevent being captured/dive
into risky situations to secure a kill, but i DON'T think it's fine to
basically guarantee a kill against a xeno/pred/whatever with zero
consequences to yourself besides waiting for recovery - nothing else in
the game incentivizes suiciding besides this item and everything else
comes with a greater degree of risk than this item because you have to
actually win the fight first for the kill and you can still be
captured/dragged/flinged away.

# Testing Photographs and Procedure
n/a, i did test on local to make sure it wouldn't gib the wrong person,
everything else works normally

# Changelog

:cl:
balance: undamaged lungemines now gib the user on hit, ex_act was
removed due to being bugged (caused runtimes, did not do anything to
xenomorphs due to checking the user for the explosive pen so ultimately
won't change anything about it)
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>