User agents for byond downloads and windows build caching (#9265)
# About the pull request

This PR is sort of a port of:
- https://github.com/BeeStation/BeeStation-Hornet/pull/12705
- https://github.com/BeeStation/BeeStation-Hornet/pull/12712
- https://github.com/tgstation/tgstation/pull/91101

Changes done:
- Docker file now uses the same url as the others (if there was a reason
this was different let me know but TG didn't have it this way)
- All curls to download byond builds now specify a useragent unique to
CM (This is for the AI overlord at cloudflare to trust the request more)
- Windows builds should attempt to cache the build (though it won't be
cached if the action is canceled - TG I guess has a different
implementation to handle this scenario but its whatever).

~~However, while Byond actually is down, I can't really see if these
changes will work so will poke at them when it looks like Byond is
available again.~~

# Testing Photographs and Procedure
<details>

- Windows build cache miss:
https://github.com/cmss13-devs/cmss13/actions/runs/14956609862/job/42013146369?pr=9265#step:4:9
- Windows build cache hit:
https://github.com/cmss13-devs/cmss13/actions/runs/14956676591/job/42013288863?pr=9265#step:4:9
- Docker file: 

![image](https://github.com/user-attachments/assets/d1ac8781-f67c-4d51-beeb-696397018581)

</details>

# Explain why it's good for the game

Passing CI!

# Changelog

No user facing changes.