make mag harnesses always prefer the slot it usually snaps to, ports sticky slots from tgmc (#8876)
# About the pull request

title
Ports tgstation/TerraGov-Marine-Corps#17104

# Explain why it's good for the game

qol
Closes #8739
Closes #9216 


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: quick equip will try to put an item to the slot it was drawn from.
qol: mag harnessed weapons always prefers the slots it usually snaps to
on quick-equip
/:cl: