Fixes M5 Camera Gear Vendor(s) (#9267)
# About the pull request

This PR takes the M5 Camera Gear fixes from #8487, I opened the branch
at the same time as the PR but was told it'd be easier to just wait for
that to be merged :D

The "fixes" is fixing that the M5 Camera Gear doesn't spawn in the squad
prep, EVER, and does NOT scale with population. There are 0 cameras at
50 pop and 0 cameras at 200 pop.

I'm completely open for the numbers to be changed.

# Explain why it's good for the game

It's good because this means SLs with no helmets, don't have to rush to
Req Line and can vend it out of their squad prep. :D


# Changelog

:cl:
add: Increased the number of M5 Camera Gears from 3 to 15 in the
Requsition Vendor.
fix: M5 Camera Gear now properly spawns in the Squad Vendors.
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>