Fix: Edit-Hotkey's Search no longer cares about case (#9276)
# About the pull request
For reasons... it wasn't lower-casing the searched for value but was
lower casing the terms being searched.
So I added the same .toLowerCase _fucking Camel Case_ to the searched
value.
This brings it into line with one another so in a sense it's
case-insensitive.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
For the longest time I just thought the search bar was broken. In a way
it was, but it was just because I started my search with a capital
letter. Suffering.
Is a fix. 
Uhhhh. 
# Testing Photographs and Procedure
<details>
<summary>Here have some screenshots</summary>

Before:


![image](https://github.com/user-attachments/assets/a058b104-bab5-4bea-9d2d-927b3687f8f2)

After:


![image](https://github.com/user-attachments/assets/4515c86e-1439-44c3-a65e-cb820496b79b)

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Edit Hotkey's search bar is case-insensitive.
/:cl: