Adds Intel SL tracking and default tracking to XO/CO headsets (#9264)
# About the pull request

As title

# Explain why it's good for the game

It's nice to be able to track the Intel SL, and it makes no sense that
the command headsets can't track anyone.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added tracking to command headsets.
add: Added a tracker for Intel Squad Lead.
add: Added XO and CO subtypes for the senior command headset. This has
no identifying features aside from different tracking options.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>